refactor(`Core.Services`): partial refactor of `MessengerService`
